backend/requirements.txt
fastapi
uvicorn
praw
vaderSentiment
transformers
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
torch
pandas